Split template headers and subdirectories
